+++ LanguageFeatures/Inline-classes/static_analysis_inline_class_A12_t01.dart
