--- LanguageFeatures/Inline-classes/static_analysis_inline_class_A18_t01.dart
@@ -1,24 +1,0 @@
-// Copyright (c) 2023, the Dart project authors.  Please see the AUTHORS file
-// for details. All rights reserved. Use of this source code is governed by a
-// BSD-style license that can be found in the LICENSE file.
-
-/// @assertion An inline type V used as an expression (a type literal) is
-/// allowed and has static type Type.
-///
-/// @description Checks that an inline type `V` can be used as an expression (a
-/// type literal) is and has static type `Type`.
-/// @author sgrekhov22@gmail.com
-
-// SharedOptions=--enable-experiment=inline-class
-
-import "../../Utils/static_type_helper.dart";
-
-inline class V {
-  final int id;
-  V(this.id);
-}
-
-main() {
-  var t = V;
-  t.expectStaticType<Exactly<Type>>();
-}
